chatter/requirements.txt
yaspin
rich==13.6.0
google-generativeai
python-dotenv
